refactor: 웹 어댑터 모듈 생성
