+++ adapter/web/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    implementation(project(":domain"))
+    implementation(project(":application"))
+
+    implementation("org.springframework.boot:spring-boot-starter-web")
+}
