+++ adapter/web/src/main/java/com/ticketlounge/web/interceptor/CurrentUserArgumentResolver.java
@@ -1,0 +1,32 @@
+package com.ticketlounge.web.interceptor;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+import com.ticketlounge.domain.auth.CurrentUser;
+import com.ticketlounge.domain.auth.MemberToken;
+
+@Component
+public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolver {
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.getParameterType().equals(MemberToken.class) &&
+                parameter.hasParameterAnnotation(CurrentUser.class);
+    }
+
+    @Override
+    public MemberToken resolveArgument(
+            MethodParameter parameter, ModelAndViewContainer mavContainer,
+            NativeWebRequest webRequest, WebDataBinderFactory binderFactory
+    ) throws Exception {
+        HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
+        return (MemberToken) request.getAttribute("memberToken");
+    }
+}
