+++ adapter/web/src/main/java/com/ticketlounge/web/ticket/request/TicketRequest.java
@@ -1,0 +1,10 @@
+package com.ticketlounge.web.ticket.request;
+
+import com.ticketlounge.application.ticket.port.in.CreateTicketUseCase.CreateTicketCommand;
+
+public record TicketRequest(String howToAcquire, int count) {
+
+    public CreateTicketCommand toCreateTicketCommand() {
+        return new CreateTicketCommand(howToAcquire, count);
+    }
+}
