+++ adapter/web/src/main/java/com/ticketlounge/web/utils/JsonException.java
@@ -1,0 +1,17 @@
+package com.ticketlounge.web.utils;
+
+public class JsonException extends RuntimeException {
+
+    public JsonException(String message) {
+        super(message);
+    }
+
+    public JsonException(Throwable cause) {
+        super(cause);
+    }
+
+    public JsonException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
