--- src/main/java/com/ticketexchange/auth/CurrentUserArgumentResolver.java
@@ -1,29 +1,0 @@
-package com.ticketexchange.auth;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolver {
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType().equals(MemberToken.class) &&
-                parameter.hasParameterAnnotation(CurrentUser.class);
-    }
-
-    @Override
-    public MemberToken resolveArgument(
-            MethodParameter parameter, ModelAndViewContainer mavContainer,
-            NativeWebRequest webRequest, WebDataBinderFactory binderFactory
-    ) throws Exception {
-        HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        return (MemberToken) request.getAttribute("memberToken");
-    }
-}
