--- src/main/java/com/ticketexchange/member/adapter/in/web/MemberController.java
@@ -1,71 +1,0 @@
-package com.ticketexchange.member.adapter.in.web;
-
-import java.net.URI;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.ticketexchange.member.adapter.in.web.request.LoginRequest;
-import com.ticketexchange.member.adapter.in.web.request.MemberRequest;
-import com.ticketexchange.member.adapter.in.web.response.LoginResponse;
-import com.ticketexchange.member.adapter.in.web.response.MemberResponse;
-import com.ticketexchange.member.application.port.in.CreateMemberUseCase;
-import com.ticketexchange.member.application.port.in.DuplicateMemberDetailsQuery;
-import com.ticketexchange.member.application.port.in.LoginMemberUseCase;
-import com.ticketexchange.support.web.ApiResult;
-
-@RestController
-@RequestMapping("/api/v1")
-public class MemberController {
-
-    private final CreateMemberUseCase createMemberUseCase;
-    private final LoginMemberUseCase loginMemberUseCase;
-    private final DuplicateMemberDetailsQuery duplicateMemberDetailsQuery;
-
-    public MemberController(
-            final CreateMemberUseCase createMemberUseCase,
-            final LoginMemberUseCase loginMemberUseCase,
-            final DuplicateMemberDetailsQuery duplicateMemberDetailsQuery
-    ) {
-        this.createMemberUseCase = createMemberUseCase;
-        this.loginMemberUseCase = loginMemberUseCase;
-        this.duplicateMemberDetailsQuery = duplicateMemberDetailsQuery;
-    }
-
-    @PostMapping("/members")
-    public ResponseEntity<ApiResult<MemberResponse>> createMember(@RequestBody MemberRequest memberRequest) {
-        MemberResponse memberResponse = MemberResponse.from(
-                createMemberUseCase.createMember(memberRequest.toCreateMemberCommand()));
-
-        return ResponseEntity.created(URI.create("/members/" + memberResponse.memberId()))
-                .body(ApiResult.succeed(memberResponse));
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<ApiResult<LoginResponse>> login(@RequestBody LoginRequest loginRequest) {
-        LoginResponse loginResponse = LoginResponse.from(loginMemberUseCase.login(loginRequest.toLoginCommand()));
-        return ResponseEntity.ok(ApiResult.succeed(loginResponse));
-    }
-
-    @GetMapping("/members/email")
-    public ResponseEntity<Void> duplicateEmail(@RequestParam String email) {
-        if (duplicateMemberDetailsQuery.isDuplicateEmail(email)) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
-        }
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @GetMapping("/members/nickname")
-    public ResponseEntity<Void> duplicateNickname(@RequestParam String nickname) {
-        if (duplicateMemberDetailsQuery.isDuplicateNickname(nickname)) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
-        }
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-}
