--- src/main/java/com/ticketexchange/member/adapter/in/web/response/LoginResponse.java
@@ -1,8 +1,0 @@
-package com.ticketexchange.member.adapter.in.web.response;
-
-public record LoginResponse(String token) {
-
-    public static LoginResponse from(String token) {
-        return new LoginResponse(token);
-    }
-}
