--- src/main/java/com/ticketexchange/utils/JsonException.java
@@ -1,17 +1,0 @@
-package com.ticketexchange.utils;
-
-public class JsonException extends RuntimeException {
-
-    public JsonException(String message) {
-        super(message);
-    }
-
-    public JsonException(Throwable cause) {
-        super(cause);
-    }
-
-    public JsonException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
